Add files via upload
Solved the thing with the strange element contents of the matrix with a bit of cleaning of the code.
First the processor initializes the 2 matrices and prints the contents of the first one. Then, it receives the awake message from the DSP. Subsequently, it send the first matrix to the DSP.
The DSP then, adds 2 to each of the matrix elements and send it back to the GPP.
Finally the GPP, receives and prints the updated by the DSP matrix.

Discarded the variable matrix size for the moment and created a function in a separate file for the initialization of the matrices.